- Huge refactoring of old Activities to new Fragment based activities (those 3 are almost done: Welcome Activity, Pie Activity and Running Processes Activity) - Externalizing styles, dimensions, attributes, strings - Debugger now includes a global log() method [Running processes area] - Ability to kill a single process using a Kill button next to its name - Better layout inflater for running processes list
git-svn-id: https://hasmoni.unfuddle.com/svn/hasmoni_rambo/trunk@45 1849bb56-f0e2-48ca-bffa-162ff6fef5c1

+++ src/com/hasmobi/rambo/utils/SingleProcess.java
@@ -8,7 +8,7 @@ public class SingleProcess {
 	public String name;
 	public boolean whitelisted = false;
 	public Drawable icon;
-	public int memoryUsage;
+	public float memoryUsage;
 	public ApplicationInfo appInfo;
 
 	public SingleProcess(String name, boolean whitelisted, Drawable icon) {
@@ -26,7 +26,7 @@ public class SingleProcess {
 	}
 
 	public SingleProcess(ApplicationInfo ai, String appName,
-			boolean isWhitelisted, Drawable appIcon, int totalMemoryUsage) {
+			boolean isWhitelisted, Drawable appIcon, float totalMemoryUsage) {
 		this.appInfo = ai;
 		this.name = appName;
 		this.whitelisted = isWhitelisted;
